refactor: extract useFlattenRecords hook.
